requirements.txt
requests
pandas
numpy
matplotlib
scikit-learn
kagglehub
pillow
torch-summary
dvc
dagshub
mlflow
dvc-s3
python-dotenv